requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
httpx
alembic
pytest